P4/requirements.txt
# P4 - Servicio de Atención Automatizada - Dependencias Integradas
# Este archivo combina las dependencias de P1, P2 y P3 con adiciones para P4

# === CORE DE P1 - DATAOPS PIPELINE ===
# Data Processing & Validation
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
great-expectations>=0.15.0
pydantic>=1.10.0

# Data Versioning & Storage
dvc>=3.0.0
boto3>=1.26.0  # AWS S3
google-cloud-storage>=2.7.0  # GCS
azure-storage-blob>=12.14.0  # Azure Blob

# Monitoring & Quality
evidently>=0.2.0
whylogs>=1.3.0
mlflow>=2.5.0
wandb>=0.15.0

# === CORE DE P2 - QA DATASET ===
# Web scraping y procesamiento
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data validation and quality
jsonschema>=4.17.0
pandera>=0.15.0

# Machine Learning and NLP
transformers>=4.30.0
datasets>=2.12.0
snorkel>=0.9.0

# Annotation
label-studio>=1.7.0

# === CORE DE P3 - AIOPS ===
# API & Web
fastapi>=0.100.0
uvicorn>=0.22.0

# Monitoreo y métricas
prometheus-client>=0.17.0

# === ADICIONES ESPECÍFICAS DE P4 ===
# Motor de búsqueda y embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0  # Búsqueda vectorial
chromadb>=0.4.0   # Vector database alternativo
langchain>=0.0.300  # Orquestación de LLMs

# Procesamiento de lenguaje mejorado
spacy>=3.6.0
nltk>=3.8.0

# Base de datos para metadatos
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# sqlite3 es parte de la librería estándar de Python, no se instala con pip [P4]

# Cache y rendimiento
redis>=4.6.0
python-memcached>=1.62

# Seguridad y autenticación
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Testing & Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
ruff>=0.5.0  # [P4] Linter usado en CI

# Utilities & Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
jinja2>=3.1.0

# Scheduling & Monitoring
celery>=5.3.0
prefect>=2.10.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Reportes y visualización
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Manejo de archivos y documentos
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
